Rename the Jetbrains app to Webview
This follows from it being used in both VSCode and JetBrains
