--- src/jetbrains/README.md
@@ -1,9 +1,0 @@
-# JetBrains Frontend
-
-This is the code rendering the graphs in the JetBrains plugin.
-
-To build it, run:
-
-```shell
-bun build-jetbrains
-```
