+++ src/webview/README.md
@@ -1,0 +1,9 @@
+# WebView Frontend
+
+This is the code rendering the graphs in the VSCode extension and the JetBrains plugin.
+
+To build it, run:
+
+```shell
+bun build-webview
+```
