+++ src/webview/public/favicon.png
