+++ src/webview/svelte.config.js
