+++ src/webview/vite.config.js
@@ -6,7 +6,7 @@ export default defineConfig({
   plugins: [svelte()],
   build: {
     copyPublicDir: true,
-    outDir: "../../dist/jetbrains",
+    outDir: "../../dist/webview",
     emptyOutDir: true,
     rollupOptions: {
       output: {
